VGAE addition minor
